pyproject.toml
[build-system]
requires = ["setuptools>=40.6.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyadi-jif"
version = "0.1.4"
description = "Python interface and configurator for ADI JESD Interface Framework"
authors = [{ name = "Travis F. Collins", email = "travis.collins@analog.com" }]
maintainers = [
    {name = "Travis Collins", email="travis.collins@analog.com"},
    {name = "Analog Devices, Inc."}
]
license = { text = "EPL-2.0" }
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
]

dependencies = [
    "numpy",
    "openpyxl>=3.1.3",
    "pandas>=1.1.5",
]

[project.optional-dependencies]
cplex = ["cplex", "docplex"]
gekko = ["gekko"]
draw = ["pyd2lang-native==0.0.3"]
tools = ["streamlit"]
e2e = [
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.4",
    "pytest-base-url>=2.0.0",
    "pytest-xdist>=3.5.0",
    "pillow>=10.0.0",
    "pixelmatch>=0.3.0",
]

[project.scripts]
jiftools = "adijif.tools.explorer.cli:run_streamlit"

[tool.coverage.paths]
source = ["adijif"]

[tool.coverage.run]
branch = true
source = ["adijif"]

[tool.coverage.report]
show_missing = true
fail_under = 90

[tool.setuptools.packages.find]
where = ["."]
include = ["adijif*"]
exclude = ["tests*"]
namespaces = true

[tool.setuptools]
py-modules = ["adijif"]

[tool.setuptools.package-data]
adijif = ["adijif/converters/resources/*", "adijif/d2/*.so", "adijif/d2/*.h"]
